# Conflicts:
#	ObligatorioP3/Alumno.cpp
#	ObligatorioP3/Asignatura.cpp
#	ObligatorioP3/Asignatura.h
#	ObligatorioP3/Carrera.h
#	ObligatorioP3/Curso.cpp
#	ObligatorioP3/Curso.h
#	ObligatorioP3/Fecha.cpp
#	ObligatorioP3/ObligatorioP3.depend
#	ObligatorioP3/ObligatorioP3.layout

+++ ObligatorioP3/Alumno.h
@@ -4,7 +4,7 @@
 #include "Escolaridad.h"
 
 typedef struct {
-    int cedula;
+    long cedula;
     String nombre;
     String apellido;
     String telefono;
